Create lambda gateway
